feat(baseline): Add new model for baseline.
This baseline comes from the admission process.

Once an applicant makes progress over the unit 1 of the admission course, her baseline data gets stored into a firestore collection called `baseline`, but it is not replicated into MongoDB.

So, The idea behind this proposal is to create a model for the baseline collection that comes from Firestore.